+++ src/main/java/baseball/domain/util/Output.java
@@ -1,4 +1,4 @@
-package baseball.domain;
+package baseball.domain.util;
 
 public class Output {
